Added MyTested.AspNetCore.Mvc.ViewData.Test (#125)
