+++ src/MyTested.AspNetCore.Mvc.ViewData/Plugins/ViewDataTestPlugin.cs
@@ -1,0 +1,18 @@
+﻿namespace MyTested.AspNetCore.Mvc.Plugins
+{
+    using System;
+    using Microsoft.Extensions.DependencyInjection;
+
+    public class ViewDataTestPlugin : IDefaultRegistrationPlugin
+    {
+        public long Priority => -1000;
+
+        public Action<IServiceCollection> DefaultServiceRegistrationDelegate
+            => serviceCollection => serviceCollection
+                .AddMvcCore()
+                .AddFormatterMappings()
+                .AddViews()
+                .AddDataAnnotations()
+                .AddJsonFormatters();
+    }
+}
