+++ src/MyTested.AspNetCore.Mvc.ViewFeatures/project.json
@@ -30,6 +30,7 @@
     "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0-*",
     "MyTested.AspNetCore.Mvc.DataAnnotations": "1.0.0-*",
     "MyTested.AspNetCore.Mvc.TempData": "1.0.0-*",
+    "MyTested.AspNetCore.Mvc.ViewData": "1.0.0-*",
     "MyTested.AspNetCore.Mvc.ViewActionResults": "1.0.0-*"
   },
