+++ src/MyTested.AspNetCore.Mvc.ViewFeatures/project.lock.json
@@ -1175,6 +1175,20 @@
         "runtime": {
           "net451/MyTested.AspNetCore.Mvc.ViewActionResults.dll": {}
         }
+      },
+      "MyTested.AspNetCore.Mvc.ViewData/1.0.0-preview-final": {
+        "type": "project",
+        "framework": ".NETFramework,Version=v4.5.1",
+        "dependencies": {
+          "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0",
+          "MyTested.AspNetCore.Mvc.Abstractions": "1.0.0"
+        },
+        "compile": {
+          "net451/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        },
+        "runtime": {
+          "net451/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        }
       }
     },
     ".NETStandard,Version=v1.6": {
@@ -3659,6 +3673,20 @@
         },
         "runtime": {
           "netstandard1.6/MyTested.AspNetCore.Mvc.ViewActionResults.dll": {}
+        }
+      },
+      "MyTested.AspNetCore.Mvc.ViewData/1.0.0-preview-final": {
+        "type": "project",
+        "framework": ".NETStandard,Version=v1.6",
+        "dependencies": {
+          "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0",
+          "MyTested.AspNetCore.Mvc.Abstractions": "1.0.0"
+        },
+        "compile": {
+          "netstandard1.6/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        },
+        "runtime": {
+          "netstandard1.6/MyTested.AspNetCore.Mvc.ViewData.dll": {}
         }
       }
     }
@@ -8192,6 +8220,11 @@
       "type": "project",
       "path": "../MyTested.AspNetCore.Mvc.ViewActionResults/project.json",
       "msbuildProject": "../MyTested.AspNetCore.Mvc.ViewActionResults/MyTested.AspNetCore.Mvc.ViewActionResults.xproj"
+    },
+    "MyTested.AspNetCore.Mvc.ViewData/1.0.0-preview-final": {
+      "type": "project",
+      "path": "../MyTested.AspNetCore.Mvc.ViewData/project.json",
+      "msbuildProject": "../MyTested.AspNetCore.Mvc.ViewData/MyTested.AspNetCore.Mvc.ViewData.xproj"
     }
   },
   "projectFileDependencyGroups": {
@@ -8199,7 +8232,8 @@
       "Microsoft.AspNetCore.Mvc.ViewFeatures >= 1.0.0-*",
       "MyTested.AspNetCore.Mvc.DataAnnotations >= 1.0.0-*",
       "MyTested.AspNetCore.Mvc.TempData >= 1.0.0-*",
-      "MyTested.AspNetCore.Mvc.ViewActionResults >= 1.0.0-*"
+      "MyTested.AspNetCore.Mvc.ViewActionResults >= 1.0.0-*",
+      "MyTested.AspNetCore.Mvc.ViewData >= 1.0.0-*"
     ],
     ".NETFramework,Version=v4.5.1": [],
     ".NETStandard,Version=v1.6": []
