+++ test/MyTested.AspNetCore.Mvc.Test/project.lock.json
@@ -3665,6 +3665,20 @@
           "netstandard1.6/MyTested.AspNetCore.Mvc.ViewActionResults.dll": {}
         }
       },
+      "MyTested.AspNetCore.Mvc.ViewData/1.0.0-preview-final": {
+        "type": "project",
+        "framework": ".NETStandard,Version=v1.6",
+        "dependencies": {
+          "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0",
+          "MyTested.AspNetCore.Mvc.Abstractions": "1.0.0"
+        },
+        "compile": {
+          "netstandard1.6/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        },
+        "runtime": {
+          "netstandard1.6/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        }
+      },
       "MyTested.AspNetCore.Mvc.ViewFeatures/1.0.0-preview-final": {
         "type": "project",
         "framework": ".NETStandard,Version=v1.6",
@@ -3672,7 +3686,8 @@
           "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0",
           "MyTested.AspNetCore.Mvc.DataAnnotations": "1.0.0",
           "MyTested.AspNetCore.Mvc.TempData": "1.0.0",
-          "MyTested.AspNetCore.Mvc.ViewActionResults": "1.0.0"
+          "MyTested.AspNetCore.Mvc.ViewActionResults": "1.0.0",
+          "MyTested.AspNetCore.Mvc.ViewData": "1.0.0"
         },
         "compile": {
           "netstandard1.6/MyTested.AspNetCore.Mvc.ViewFeatures.dll": {}
@@ -5386,6 +5401,20 @@
           "net451/MyTested.AspNetCore.Mvc.ViewActionResults.dll": {}
         }
       },
+      "MyTested.AspNetCore.Mvc.ViewData/1.0.0-preview-final": {
+        "type": "project",
+        "framework": ".NETFramework,Version=v4.5.1",
+        "dependencies": {
+          "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0",
+          "MyTested.AspNetCore.Mvc.Abstractions": "1.0.0"
+        },
+        "compile": {
+          "net451/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        },
+        "runtime": {
+          "net451/MyTested.AspNetCore.Mvc.ViewData.dll": {}
+        }
+      },
       "MyTested.AspNetCore.Mvc.ViewFeatures/1.0.0-preview-final": {
         "type": "project",
         "framework": ".NETFramework,Version=v4.5.1",
@@ -5393,7 +5422,8 @@
           "Microsoft.AspNetCore.Mvc.ViewFeatures": "1.0.0",
           "MyTested.AspNetCore.Mvc.DataAnnotations": "1.0.0",
           "MyTested.AspNetCore.Mvc.TempData": "1.0.0",
-          "MyTested.AspNetCore.Mvc.ViewActionResults": "1.0.0"
+          "MyTested.AspNetCore.Mvc.ViewActionResults": "1.0.0",
+          "MyTested.AspNetCore.Mvc.ViewData": "1.0.0"
         },
         "compile": {
           "net451/MyTested.AspNetCore.Mvc.ViewFeatures.dll": {}
@@ -11366,6 +11396,11 @@
       "type": "project",
       "path": "../../src/MyTested.AspNetCore.Mvc.ViewActionResults/project.json",
       "msbuildProject": "../../src/MyTested.AspNetCore.Mvc.ViewActionResults/MyTested.AspNetCore.Mvc.ViewActionResults.xproj"
+    },
+    "MyTested.AspNetCore.Mvc.ViewData/1.0.0-preview-final": {
+      "type": "project",
+      "path": "../../src/MyTested.AspNetCore.Mvc.ViewData/project.json",
+      "msbuildProject": "../../src/MyTested.AspNetCore.Mvc.ViewData/MyTested.AspNetCore.Mvc.ViewData.xproj"
     },
     "MyTested.AspNetCore.Mvc.ViewFeatures/1.0.0-preview-final": {
       "type": "project",
