+++ test/MyTested.AspNetCore.Mvc.ViewData.Test/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+using Xunit;
+
+[assembly: AssemblyProduct("MyTested.AspNetCore.Mvc.ViewData.Test")]
+[assembly: ComVisible(false)]
+
+[assembly: CollectionBehavior(DisableTestParallelization = true)]
