+++ test/MyTested.AspNetCore.Mvc.ViewData.Test/project.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "dotnet-test-xunit": "2.2.0-*",
     "xunit": "2.2.0-*",
-    "MyTested.AspNetCore.Mvc.TempData": "1.0.0-*",
+    "MyTested.AspNetCore.Mvc.Controllers": "1.0.0-*",
+    "MyTested.AspNetCore.Mvc.ViewData": "1.0.0-*",
     "MyTested.AspNetCore.Mvc.Test.Setups": "1.0.0-*"
   },
